updated affected rows to int64

+++ Sources/MySQL/MySQLResultProtocol.swift
@@ -2,6 +2,6 @@ import Foundation
 
 public protocol MySQLResultProtocol {
     var fields: [MySQLField] { get }
-    var affectedRows: Int { get }
+    var affectedRows: Int64 { get }
     func nextResult() -> MySQLRow?
 }
